Centralize name generation in one place
- Both to get more code out of objcgenerator*.cs and pave the way to possible future caching
