+++ objcgen/objcgen.csproj
@@ -320,6 +320,7 @@
     <Compile Include="sourcewriter.cs" />
     <Compile Include="classhelper.cs" />
     <Compile Include="protocolhelper.cs" />
+    <Compile Include="NameGenerator.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="support\" />
